% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clone.R
\name{clone}
\alias{clone}
\alias{clone_attributes}
\title{Create a Copy of an oClass Instance}
\usage{
clone(x, deep = FALSE, ...)

clone_attributes(x, deep = FALSE, cloned = NULL)
}
\arguments{
\item{x}{environment of class \code{"Instance"}}

\item{deep}{logical for whether clone should be applied to all objects.}

\item{...}{arguments passed to methods}

\item{cloned}{environment containing references to environments that have already been cloned. This is passed to internal methods when \code{deep=TRUE} and should not be set directly.}
}
\value{
environment of class \code{"Instance"}
}
\description{
A copy of all objects and attributes within an environment. If \code{deep=TRUE},
all objects inside of \code{x}, including other environments, will also be deeply "cloned".
The global and base environments will not be cloned.
}
\section{Functions}{
\itemize{
\item \code{clone_attributes}: Clone the attributes of an object.
}}

