% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init.Instance}
\alias{init.Instance}
\title{Init Method for Instance}
\usage{
\method{init}{Instance}(x, ...)
}
\arguments{
\item{x}{environment of class \code{"Instance"}}

\item{...}{named fields inherited by the class instance or passed to methods}
}
\value{
environment of class \code{"Instance"}
}
\description{
See \code{\link{init}} for details.
}
