% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{is.Instance}
\alias{is.Instance}
\title{Is Object a Class Instance?}
\usage{
is.Instance(x)
}
\arguments{
\item{x}{object to be tested}
}
\value{
\code{TRUE} if object inherits \code{"Instance"}, \code{FALSE} otherwise
}
\description{
Check whether object inherits the \code{"Instance"} class. See \code{\link{is.oClass}} to check whether object
is a \code{\link{oClass}} generator.
}
