% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{is.oClass}
\alias{is.oClass}
\alias{is.ClassGenerator}
\title{Is Object an "oClass" Generator?}
\usage{
is.oClass(x)

is.ClassGenerator(x)
}
\arguments{
\item{x}{object to be tested}
}
\value{
\code{TRUE} if object inherits \code{"ClassGenerator"}, \code{FALSE} otherwise
}
\description{
Check whether object inherits the \code{"ClassGenerator"} class. This is used to check \code{\link{oClass}}
generators, not the instance. See \code{\link{is.Instance}} to check whether object is an oClass instance.
}
\section{Functions}{
\itemize{
\item \code{is.ClassGenerator}: check whether object is an oClass generator
}}

