% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logoops2}
\alias{logoops2}
\alias{logoops}
\title{Runs Expectation Maximization OOPS and reestimates model parameters.}
\usage{
logoops(fasta, alpha, beta, cutoff = 0.01, niter = 20L)
}
\arguments{
\item{fasta}{Dataset of sequences.}

\item{alpha}{PWM model to be reestimated.}

\item{beta}{Markov model thats represents the control senquences.}

\item{cutoff}{Cutoff for EM convergence.}

\item{niter}{Maximum number of iterations.}

\item{priori}{Priori value per each position.}
}
\description{
Runs Expectation Maximization OOPS and reestimates model parameters.
}
